+++ stdb/data_source.py
@@ -40,7 +40,7 @@ class LocalDataSource(object):
             ('high', 'float64'), ('low', 'float64'),
             ('close', 'float64'), ('volume', 'float64'),
             ('vwap', 'float64'), ('returns', 'float64'),
-            ('amount','uint64')
+            ('amount','float64')
         ])
         bars = np.array(history_data,dtype=stocktype)
         bars = bars[::-1]#转向
